'long double' high-precision search version (slower)

+++ polapt-test_y.h
@@ -16,7 +16,7 @@
  */
 
 /* -0.0, *1.0 */
-static inline TEST_T test_sqrt_r1_d4(TEST_T x, double scale_adj[]) {
+static inline TEST_T test_sqrt_r1_d4(TEST_T x, long double scale_adj[]) {
 	const TEST_T scale[] = {
 		+24.344885f/6 * scale_adj[0],
 		-58.344885f/6 * scale_adj[1],
@@ -24,12 +24,12 @@ static inline TEST_T test_sqrt_r1_d4(TEST_T x, double scale_adj[]) {
 		-27.344885f/6 * scale_adj[3],
 	};
 	TEST_T x2 = x*x;
-	TEST_T xa = fabs(x);
+	TEST_T xa = fabsl(x);
 	return x*(scale[0] + xa*(scale[1] + xa*(scale[2] + xa*scale[3])));
 }
 
 /* 0.0, *1.0 */
-static inline TEST_T test_cosramp_jkp(TEST_T x, double scale_adj[]) {
+static inline TEST_T test_cosramp_jkp(TEST_T x, long double scale_adj[]) {
 	const TEST_T scale[] = {
 		+118.f/75 * scale_adj[0],
 		-196.f/75 * scale_adj[1],
@@ -50,26 +50,26 @@ static inline TEST_T test_cosramp_jkp(TEST_T x, double scale_adj[]) {
 }
 
 /* -0.5, *2.0 */
-static inline TEST_T test_fabs_d16(TEST_T x, double scale_adj[]) {
+static inline TEST_T test_fabs_d16(TEST_T x, long double scale_adj[]) {
 	const TEST_T scale[] = {
 		+1.f * scale_adj[0],
 		-1.f * scale_adj[1],
 		+1.f * scale_adj[2],
 		-1.f * scale_adj[3],
 	};
-	TEST_T xp = (x+x - x*fabs(x));
+	TEST_T xp = (x+x - x*fabsl(x));
 	xp *= xp;
 	return xp*(scale[0] + xp*(scale[1] + xp*(scale[2] + xp*scale[3])));
 }
 
 /* -0.5, *2.0 */
-static inline TEST_T test_fabs_d12(TEST_T x, double scale_adj[]) {
+static inline TEST_T test_fabs_d12(TEST_T x, long double scale_adj[]) {
 	const TEST_T scale[] = {
 		+1.f * scale_adj[0],
 		-1.f * scale_adj[1],
 		+1.f * scale_adj[2],
 		-1.f * scale_adj[3],
 	};
-	TEST_T xp = (3 - 2*fabs(x))*x*x;
+	TEST_T xp = (3 - 2*fabsl(x))*x*x;
 	return xp*(scale[0] + xp*(scale[1] + xp*(scale[2] + xp*scale[3])));
 }
